Added admin controls and changed loan sanction method

+++ main.py
@@ -1,8 +1,11 @@
 import mysql.connector
-from flask import Flask, request, render_template, redirect
+from flask import Flask, request, render_template, redirect, session
 import time
 from datetime import datetime
+
 datetime.today().strftime('%Y-%m-%d')
+
+
 
 
 try :
@@ -21,7 +24,7 @@ except:
 
 
 app = Flask('__name__')
-
+app.secret_key = 'iownthisbank'
 
 # Login Page
 
@@ -34,14 +37,15 @@ def login():
 @app.route("/", methods = ['POST','GET'])
 def loginverify():
 
-
     if request.method == 'POST':
         accno = request.form["accno"]
         password = request.form["password"]
         try : 
             curr.execute('select passwordCheck(%s,%s)', [int(accno), password])
             validity = curr.fetchone()
+            print(validity)
             if validity[0] == 1:
+                session['useraccno'] = int(accno)
                 return redirect('/' + str(accno))
             else:
                 return render_template('login.html', valid = 'Please re-check the credentials')
@@ -53,6 +57,9 @@ def loginverify():
 
 @app.route('/<int:accno>')
 def userdetails(accno):
+    if session['useraccno'] != accno:
+        return render_template('login.html', valid = 'Re-login to continue')
+
     curr.execute('select balance, min_balance from balance_details where accno = %s', [int(accno)])
     bldetails = curr.fetchone()
     curr.execute('select * from transactions where sender_accno = %s', [int(accno)])
@@ -63,7 +70,6 @@ def userdetails(accno):
     loans = curr.fetchall()
     curr.execute('select name from accounts where accno = %s', [int(accno)])
     name = curr.fetchone()
-    print(bldetails)
 
     return render_template('userdetails.html', bldetails = bldetails, transactions=transactions, investments=investments, loans=loans,accno=accno, name = name[0])
 
@@ -71,6 +77,8 @@ def userdetails(accno):
 
 @app.route('/<int:accno>transact')
 def transactPage(accno):
+    if session['useraccno'] != accno:
+        return render_template('login.html', valid = 'Re-login to continue')
     return render_template('transact.html', accno=accno)
 
 
@@ -78,6 +86,8 @@ def transactPage(accno):
 
 @app.route('/<int:accno>transact', methods=['POST'])
 def transact(accno):
+    if session['useraccno'] != accno:
+        return render_template('login.html', valid = 'Re-login to continue')
     if request.method == 'POST':
         raccno = request.form["raccno"]
         amount = request.form["amount"]
@@ -96,10 +106,14 @@ def transact(accno):
 
 @app.route('/<int:accno>openinvestment')
 def investmentpage(accno):
+    if session['useraccno'] != accno:
+        return render_template('login.html', valid = 'Re-login to continue')
     return render_template('investment.html', accno = accno, status = '')
 
 @app.route('/<int:accno>openinvestment', methods=['POST'])
 def investmentopen(accno):
+    if session['useraccno'] != accno:
+        return render_template('login.html', valid = 'Re-login to continue')
     if request.method == 'POST':
         amount = request.form["amount"]
         mdate = request.form["mdate"]
@@ -122,6 +136,8 @@ def investmentopen(accno):
 # Closing Investment
 @app.route('/<int:accno>+<int:investmentid>investment', methods = ['POST'])
 def closeinvestment(accno, investmentid):
+    if session['useraccno'] != accno:
+        return render_template('login.html', valid = 'Re-login to continue')
     if request.method == 'POST':
         curr.execute('delete from investments where investment_id = %s', [investmentid])
         curr.execute('commit')
@@ -130,6 +146,8 @@ def closeinvestment(accno, investmentid):
 # Closing Loan
 @app.route('/<int:accno>+<int:loanid>loan', methods=['POST'])
 def repayLoan(accno, loanid):
+    if session['useraccno'] != accno:
+        return render_template('login.html', valid = 'Re-login to continue')
     if request.method == 'POST':
         try:
             curr.execute('select repayLoan(%s,%s)', [accno, loanid])
@@ -147,17 +165,21 @@ def repayLoan(accno, loanid):
 
 @app.route('/<int:accno>loan')
 def loan(accno):
+    if session['useraccno'] != accno:
+        return render_template('login.html', valid = 'Re-login to continue')
     return render_template('loan.html', accno=accno)
 
 #Submit Loan
 @app.route('/<int:accno>loan', methods=["POST"])
 def loansanction(accno):
+    if session['useraccno'] != accno:
+        return render_template('login.html', valid = 'Re-login to continue')
     amount = request.form["lamount"]
     rdate = request.form["rdate"]
     cdate = datetime.today().strftime('%Y-%m-%d')
 
     try:
-        curr.execute('insert into loans values(null,%s,%s,null,%s,%s,null,null)', [accno, int(amount),cdate,rdate])
+        curr.execute('insert into loans values(null,%s,%s,null,%s,%s,null,null,null)', [accno, int(amount),cdate,rdate])
         curr.execute('commit')
         # redirect('/' + str(accno))
 
@@ -165,6 +187,67 @@ def loansanction(accno):
         print(err)
 
     return redirect('/' + str(accno))
+
+
+#Admin Login
+@app.route('/adminlogin')
+def admin():
+    return render_template('adminlogin.html')
+
+
+#Admin verification
+@app.route('/adminlogin', methods = ["POST"])
+def adminverify():
+    if request.method == "POST":
+        print(request.form["password"])
+        print(app.secret_key)
+        if app.secret_key == request.form["password"]:
+            session['adminlogin'] = 'verified'
+            return redirect('/adminlogin/admincontrols')
+
+    return redirect('/adminlogin')
+
+@app.route('/adminlogin/admincontrols')
+def admincontrols():
+    if session['adminlogin'] == 'verified':
+        curr.execute('select * from accounts')
+        accounts = curr.fetchall()
+        curr.execute('select * from balance_details')
+        bldetails = curr.fetchall()
+        curr.execute('select * from transactions where sender_accno')
+        transactions = curr.fetchall()
+        curr.execute('select * from investments')
+        investments = curr.fetchall()
+        curr.execute("select * from loans")
+        loans = curr.fetchall()
+        
+        return render_template('admincontrols.html', accounts = accounts, transactions = transactions, loans = loans, investments = investments, balances = bldetails)
+    
+    return redirect('/adminlogin')
+
+#Balance Updation
+@app.route('/adminlogin/<int:accno>+balanceupdate', methods=["POST"])
+def updatebalance(accno):
+    if session['adminlogin'] == 'verified':
+        if request.method == "POST":
+            amount = request.form["amount"]
+            try :
+                curr.execute("update balance_details set balance = balance + %s where accno = %s", [int(amount), int(accno)])
+                curr.execute("commit")
+            except mysql.connector.Error as err:
+                print(err)
+    return redirect('/adminlogin/admincontrols')
+
+#Loan Sanction
+@app.route('/adminlogin/<int:accno>+<int:loanid>+<int:amount>sanctionloan', methods=['POST'])
+def sanctionloan(accno, loanid, amount):
+    if session['adminlogin'] == 'verified':
+        if request.method == 'POST':
+            curr.execute('select sanctionLoan(%s,%s,%s)',[int(accno), int(loanid), int(amount)])
+            curr.fetchall()
+            curr.execute('commit')
+    return redirect('/adminlogin/admincontrols')
+
 
 
 if __name__ == '__main__':
